needs some chandes

+++ controllers/feed.js
@@ -120,6 +120,11 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 404
 				throw Error
 			}
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized')
+				error.statusCode = 403
+				throw error
+			}
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl)
 			}
@@ -148,7 +153,13 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404
 				throw error
 			}
-			//! CHECKING IF THE USER IS A CREATOR (NEED TO ADD)
+			//! CHECKING IF THE USER IS A CREATOR
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized')
+				error.statusCode = 403
+				throw error
+			}
+
 			clearImage(post.imageUrl)
 			return Post.findByIdAndDelete(postId)
 		})
